Two type mismatch fixes previously committed to trunk\ and forgotten from backports
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.2.x@605073 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/http/http_etag.c
@@ -37,7 +37,8 @@ static char *etag_uint64_to_hex(char *next, apr_uint64_t u)
     int printing = 0;
     int shift = sizeof(apr_uint64_t) * 8 - 4;
     do {
-        unsigned short next_digit = ((u >> shift) & (apr_uint64_t)0xf);
+        unsigned short next_digit = (unsigned short)
+                                    ((u >> shift) & (apr_uint64_t)0xf);
         if (next_digit) {
             *next++ = HEX_DIGITS[next_digit];
             printing = 1;
